Refactor hooks file into a few subfiles
Still a couple of hooks left at the top level that didn't have a clear
categorical home.
